pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "energy-forecasting"
version = "1.0.0"
description = "Multi-horizon probabilistic energy forecasting with PatchTST, TFT, and N-BEATS"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "requests>=2.31",
]

[tool.setuptools.packages.find]
where = ["src"]
